refactor settings components
